+++ CHANGES.md
@@ -9,6 +9,9 @@ Release Notes.
 * Add layer field to event when reporting.
 * Remove redundant `shade.package` property.
 * Add servicecomb-2.x plugin and Testcase.
+* Fix NPE in gateway plugin when the timer triggers webflux webclient call.
+* Add an optional plugin, trace-sampler-cpu-policy-plugin, which could disable trace collecting in high CPU load.
+* Change the dateformat of logs to `yyyy-MM-dd HH:mm:ss.SSS`(was `yyyy-MM-dd HH:mm:ss:SSS`).
 
 #### Documentation